fixed profile header
