+++ iosApp/iosApp/UI/Page/ProfileNew/View/ProfileRefreshViewController.swift
@@ -80,22 +80,31 @@ class ProfileNewRefreshViewController: UIViewController {
         navigationBar.isHidden = true  // 始终隐藏自定义导航栏
         navigationBar.alpha = 0
         
-        // 如果是自己的 profile，显示系统导航栏，隐藏返回按钮
+        // 如果是自己的 profile，显示系统导航栏，隐藏返回按钮和更多按钮
         if isOwnProfile {
             navigationController?.setNavigationBarHidden(false, animated: false)
             segmentedBackButton.isHidden = true
+            navigationBar.items?.first?.rightBarButtonItem = nil
         } else {
-               // 默认显示系统导航栏
-               // 每次切换tab后，上级就刷新，然后 这边就又有问题了，无解， 此bug 得解决上层刷新问题。
-        navigationController?.setNavigationBarHidden(false, animated: false)
-        
-            // 如果是其他用户的 profile，初始时隐藏系统导航栏和返回按钮
-            // navigationController?.setNavigationBarHidden(true, animated: false)
-            segmentedBackButton.isHidden = true
+            // 如果是其他用户的 profile，显示更多按钮
+            navigationController?.setNavigationBarHidden(false, animated: false)
+            let moreButton = UIBarButtonItem(image: UIImage(systemName: "ellipsis.circle"), style: .plain, target: self, action: #selector(handleMoreMenuTap))
+            navigationController?.navigationBar.topItem?.rightBarButtonItem = moreButton
         }
-    
+        
         // 更新UI
         updateUI()
+        
+        // 配置头部视图
+        if let userInfo = userInfo {
+            userHeaderView?.configure(with: userInfo, state: state)
+            
+            // 设置关注按钮回调
+            userHeaderView?.onFollowClick = { [weak self] relation in
+                os_log("[📔][ProfileRefreshViewController]点击关注按钮: userKey=%{public}@", log: .default, type: .debug, userInfo.profile.key.description)
+                state.follow(userKey: userInfo.profile.key, data: relation)
+            }
+        }
     }
 
     private func updateUI() {
@@ -145,6 +154,7 @@ class ProfileNewRefreshViewController: UIViewController {
         userHeaderView = ProfileNewHeaderView(frame: CGRect(x: 0, y: 0, width: UIScreen.main.bounds.width, height: 0))
         
         // 设置头部视图的回调
+        /* 注释掉原有代码，保留作为参考
         userHeaderView.onAvatarTap = { [weak self] in
             // TODO: 处理头像点击
         }
@@ -207,6 +217,12 @@ class ProfileNewRefreshViewController: UIViewController {
             
             // 显示菜单
             self.present(alertController, animated: true)
+        }
+        */
+        
+        // 新的配置代码
+        if let userInfo = userInfo {
+            userHeaderView?.configure(with: userInfo, state: state)
         }
         
         // 配置分段控制器
@@ -321,6 +337,11 @@ class ProfileNewRefreshViewController: UIViewController {
         let navigationItem = UINavigationItem()
         let backButton = UIBarButtonItem(image: UIImage(systemName: "chevron.left"), style: .plain, target: self, action: #selector(backButtonTapped))
         navigationItem.leftBarButtonItem = backButton
+        
+        // 添加更多按钮到导航栏右侧
+        let moreButton = UIBarButtonItem(image: UIImage(systemName: "ellipsis.circle"), style: .plain, target: self, action: #selector(handleMoreMenuTap))
+        navigationItem.rightBarButtonItem = moreButton
+        
         navigationBar.items = [navigationItem]
         
         // 添加到视图
@@ -328,6 +349,10 @@ class ProfileNewRefreshViewController: UIViewController {
         
         // 设置返回按钮事件
         segmentedBackButton.addTarget(self, action: #selector(backButtonTapped), for: .touchUpInside)
+        
+        // 初始时隐藏 moreButton
+        moreButton.isEnabled = false
+        navigationItem.rightBarButtonItem = nil
     }
     
     @objc private func backButtonTapped() {
@@ -337,6 +362,81 @@ class ProfileNewRefreshViewController: UIViewController {
         } else {
             self.dismiss(animated: true)
         }
+    }
+    
+    @objc private func handleMoreMenuTap() {
+        os_log("[ProfileRefreshViewController] More menu button tapped", log: .default, type: .debug)
+        
+        guard let state = state,
+              case .success(let isMe) = onEnum(of: state.isMe),
+              !isMe.data.boolValue else { return }
+        
+        // 创建更多菜单
+        let alertController = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
+        
+        // 添加屏蔽/取消屏蔽选项
+        if case .success(let relation) = onEnum(of: state.relationState),
+           case .success(let actions) = onEnum(of: state.actions),
+           actions.data.size > 0 {
+            for index in 0..<actions.data.size {
+                let item = actions.data.get(index: index)
+                let title = switch onEnum(of: item) {
+                case .block(let block): block.relationState(relation: relation.data) ?
+                    NSLocalizedString("unblock", comment: "") :
+                    NSLocalizedString("block", comment: "")
+                case .mute(let mute): mute.relationState(relation: relation.data) ?
+                    NSLocalizedString("unmute", comment: "") :
+                    NSLocalizedString("mute", comment: "")
+                }
+                
+                alertController.addAction(UIAlertAction(title: title, style: .default) { [weak self] _ in
+                    if case .success(let user) = onEnum(of: state.userState) {
+                        Task {
+                            try? await item.invoke(userKey: user.data.key, relation: relation.data)
+                            // 显示操作成功的 Toast
+                            await MainActor.run {
+                                if let window = self?.view.window {
+                                    let (icon, message) = switch onEnum(of: item) {
+                                    case .block(let block):
+                                        if block.relationState(relation: relation.data) {
+                                            (UIImage(systemName: "checkmark.circle"), NSLocalizedString("user_unblock", comment: ""))
+                                        } else {
+                                            (UIImage(systemName: "checkmark.circle"), NSLocalizedString("user_block", comment: ""))
+                                        }
+                                    case .mute(let mute):
+                                        if mute.relationState(relation: relation.data) {
+                                            (UIImage(systemName: "checkmark.circle"), "Unmuted")
+                                        } else {
+                                            (UIImage(systemName: "checkmark.circle"), "Muted")
+                                        }
+                                    }
+                                    let toastView = ToastView(icon: icon, message: message)
+                                    toastView.show(in: window)
+                                }
+                            }
+                        }
+                    }
+                })
+            }
+        }
+        
+        // 添加举报选项
+        if case .success(let user) = onEnum(of: state.userState) {
+            alertController.addAction(UIAlertAction(title: NSLocalizedString("report", comment: ""), style: .destructive) { [weak self] _ in
+                state.report(userKey: user.data.key)
+                // 显示举报成功的 Toast
+                if let window = self?.view.window {
+                    let toastView = ToastView(icon: UIImage(systemName: "checkmark.circle"), message: NSLocalizedString("report", comment: ""))
+                    toastView.show(in: window)
+                }
+            })
+        }
+        
+        // 添加取消选项
+        alertController.addAction(UIAlertAction(title: NSLocalizedString("cancel", comment: ""), style: .cancel))
+        
+        // 显示菜单
+        present(alertController, animated: true)
     }
     
     private func updateNavigationBarVisibility(with offset: CGFloat) {
@@ -531,441 +631,3 @@ extension ProfileNewRefreshViewController: JXSegmentedViewDelegate {
     }
 }
 
-// 头部视图
-class ProfileNewHeaderView: UIView {
-    private let bannerImageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.contentMode = .scaleAspectFill
-        imageView.clipsToBounds = true
-        return imageView
-    }()
-    
-    private let blurEffectView: UIVisualEffectView = {
-        let blurEffect = UIBlurEffect(style: .light)
-        let view = UIVisualEffectView(effect: blurEffect)
-        view.alpha = 0 // 初始时不模糊
-        return view
-    }()
-    
-    private let avatarView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.backgroundColor = .gray.withAlphaComponent(0.3)
-        imageView.layer.cornerRadius = 40
-        imageView.clipsToBounds = true
-        imageView.contentMode = .scaleAspectFill
-        return imageView
-    }()
-    
-    private let followButton: UIButton = {
-        let button = UIButton(type: .system)
-        button.setTitle("关注", for: .normal)
-        button.setTitleColor(.white, for: .normal)
-        button.backgroundColor = .systemBlue
-        button.layer.cornerRadius = 15
-        return button
-    }()
-    
-    private let nameLabel: UILabel = {
-        let label = UILabel()
-        label.font = .boldSystemFont(ofSize: 20)
-        return label
-    }()
-    
-    private let handleLabel: UILabel = {
-        let label = UILabel()
-        label.textColor = .gray
-        label.font = .systemFont(ofSize: 15)
-        return label
-    }()
-    
-    private let descriptionLabel: UILabel = {
-        let label = UILabel()
-        label.numberOfLines = 0
-        label.font = .systemFont(ofSize: 15)
-        return label
-    }()
-    
-    private let followsCountLabel: UILabel = {
-        let label = UILabel()
-        label.font = .systemFont(ofSize: 14)
-        label.textColor = .gray
-        return label
-    }()
-    
-    private let fansCountLabel: UILabel = {
-        let label = UILabel()
-        label.font = .systemFont(ofSize: 14)
-        label.textColor = .gray
-        return label
-    }()
-    
-    private let markStackView: UIStackView = {
-        let stackView = UIStackView()
-        stackView.axis = .horizontal
-        stackView.spacing = 4
-        stackView.alignment = .center
-        return stackView
-    }()
-    
-    private let moreButton: UIButton = {
-        let button = UIButton(type: .system)
-        button.setImage(UIImage(systemName: "ellipsis.circle"), for: .normal)
-        button.isHidden = true // 默认隐藏
-        return button
-    }()
-    
-    var onFollowsCountTap: (() -> Void)?
-    var onFansCountTap: (() -> Void)?
-    var onAvatarTap: (() -> Void)?
-    var onBannerTap: (() -> Void)?
-    var onMoreMenuTap: (() -> Void)?
-    
-  
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupUI()
-      
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
- 
-    
-    private func setupUI() {
-        backgroundColor = .systemBackground
-        
-        // Banner with tap gesture
-        addSubview(bannerImageView)
-        bannerImageView.frame = CGRect(x: 0, y: 0, width: frame.width, height: 150)
-        let bannerTap = UITapGestureRecognizer(target: self, action: #selector(bannerTapped))
-        bannerImageView.addGestureRecognizer(bannerTap)
-        bannerImageView.isUserInteractionEnabled = true
-        
-        // Blur effect
-        addSubview(blurEffectView)
-        blurEffectView.frame = bannerImageView.frame
-        
-        // Avatar with tap gesture
-        addSubview(avatarView)
-        avatarView.frame = CGRect(x: 16, y: 110, width: 80, height: 80)
-        let avatarTap = UITapGestureRecognizer(target: self, action: #selector(avatarTapped))
-        avatarView.addGestureRecognizer(avatarTap)
-        avatarView.isUserInteractionEnabled = true
-        
-        // Follow Button
-        addSubview(followButton)
-        followButton.frame = CGRect(x: frame.width - 100, y: 160, width: 80, height: 30)
-        
-        // More Menu Button
-        moreButton.addTarget(self, action: #selector(moreMenuTapped), for: .touchUpInside)
-        addSubview(moreButton)
-        moreButton.frame = CGRect(x: frame.width - 44, y: 60, width: 30, height: 30)
-        
-        // Name Label
-        addSubview(nameLabel)
-        nameLabel.frame = CGRect(x: 16, y: avatarView.frame.maxY + 10, width: frame.width - 32, height: 24)
-        
-        // Handle Label and Mark Stack
-        addSubview(handleLabel)
-        handleLabel.frame = CGRect(x: 16, y: nameLabel.frame.maxY + 4, width: frame.width - 32, height: 20)
-        
-        addSubview(markStackView)
-        markStackView.frame = CGRect(x: handleLabel.frame.maxX + 4, y: nameLabel.frame.maxY + 4, width: 100, height: 20)
-        
-        // Follows/Fans Count with tap gesture
-        addSubview(followsCountLabel)
-        followsCountLabel.frame = CGRect(x: 16, y: handleLabel.frame.maxY + 6, width: 100, height: 20)
-        let followsTap = UITapGestureRecognizer(target: self, action: #selector(followsCountTapped))
-        followsCountLabel.addGestureRecognizer(followsTap)
-        followsCountLabel.isUserInteractionEnabled = true
-        
-        addSubview(fansCountLabel)
-        fansCountLabel.frame = CGRect(x: 120, y: handleLabel.frame.maxY + 6, width: 100, height: 20)
-        let fansTap = UITapGestureRecognizer(target: self, action: #selector(fansCountTapped))
-        fansCountLabel.addGestureRecognizer(fansTap)
-        fansCountLabel.isUserInteractionEnabled = true
-        
-        // Description Label
-        addSubview(descriptionLabel)
-        descriptionLabel.frame = CGRect(x: 16, y: followsCountLabel.frame.maxY + 10, width: frame.width - 32, height: 0)
-    }
-    
-    private func layoutContent() {
-        // 计算description的高度
-        let descriptionWidth = frame.width - 32
-        let descriptionSize = descriptionLabel.sizeThatFits(CGSize(width: descriptionWidth, height: .greatestFiniteMagnitude))
-        
-        // 更新description的frame
-        descriptionLabel.frame = CGRect(x: 16, y: 280, width: descriptionWidth, height: descriptionSize.height)
-        
-        // 获取最后一个子视图的底部位置
-        var maxY: CGFloat = 0
-        for subview in subviews {
-            let subviewBottom = subview.frame.maxY
-            if subviewBottom > maxY {
-                maxY = subviewBottom
-            }
-        }
-        
-        // 更新整体高度，添加底部padding
-        frame.size.height = maxY + 16 // 16是底部padding
-    }
-    
-    
-    
-    // 更新Banner拉伸效果
-    func updateBannerStretch(withOffset offset: CGFloat) {
-        let normalHeight: CGFloat = 150
-        let stretchedHeight = normalHeight + max(0, offset)
-        
-        // 更新Banner图片frame
-        bannerImageView.frame = CGRect(x: 0, y: min(0, -offset), width: frame.width, height: stretchedHeight)
-        blurEffectView.frame = bannerImageView.frame
-        
-        // 根据拉伸程度设置模糊效果
-        let blurAlpha = min(offset / 100, 0.3)  // 最大模糊度0.3
-        blurEffectView.alpha = blurAlpha
-    }
-    
-    func getContentHeight() -> CGFloat {
-        return frame.height
-    }
-    
-    func configure(with userInfo: ProfileUserInfo) {
-        // 设置用户名
-        nameLabel.text = userInfo.profile.name.markdown
-        
-        // 设置用户handle
-        handleLabel.text = "\(userInfo.profile.handle)"
-        
-        // 设置头像 - 使用 Kingfisher 缓存
-        if let url = URL(string: userInfo.profile.avatar) {
-            avatarView.kf.setImage(
-                with: url,
-                options: [
-                    .transition(.fade(0.25)),
-                    .processor(DownsamplingImageProcessor(size: CGSize(width: 160, height: 160))),
-                    .scaleFactor(UIScreen.main.scale),
-                    .cacheOriginalImage
-                ]
-            )
-        }
-        
-        // 设置banner - 使用 Kingfisher 缓存
-        if let url = URL(string: userInfo.profile.banner ?? ""),
-           !(userInfo.profile.banner ?? "").isEmpty,
-           (userInfo.profile.banner ?? "").range(of: "^https?://.*example\\.com.*$", options: .regularExpression) == nil {
-            bannerImageView.kf.setImage(
-                with: url,
-                options: [
-                    .transition(.fade(0.25)),
-                    .processor(DownsamplingImageProcessor(size: CGSize(width: UIScreen.main.bounds.width * 2, height: 300))),
-                    .scaleFactor(UIScreen.main.scale),
-                    .cacheOriginalImage
-                ]
-            ) { result in
-                switch result {
-                case .success(let imageResult):
-                    // 检查图片是否有效
-                    if imageResult.image.size.width > 10 && imageResult.image.size.height > 10 {
-                        // 图片有效，保持现状
-                    } else {
-                        // 如果图片无效，使用头像作为背景
-                        self.setupDynamicBannerBackground(avatarUrl: userInfo.profile.avatar)
-                    }
-                case .failure(_):
-                    // 加载失败，使用头像作为背景
-                    self.setupDynamicBannerBackground(avatarUrl: userInfo.profile.avatar)
-                }
-            }
-        } else {
-            // 如果没有banner，使用头像作为背景
-            setupDynamicBannerBackground(avatarUrl: userInfo.profile.avatar)
-        }
-        
-        // 设置关注/粉丝数
-        followsCountLabel.text = "\(userInfo.followCount) 关注"
-        fansCountLabel.text = "\(userInfo.fansCount) 粉丝"
-        
-        // 更新关注按钮状态
-        updateFollowButton(with: userInfo)
-        
-        // 设置用户标记
-        markStackView.arrangedSubviews.forEach { $0.removeFromSuperview() }
-        for mark in userInfo.profile.mark {
-            let imageView = UIImageView()
-            imageView.tintColor = .gray
-            imageView.alpha = 0.6
-            
-            switch mark {
-            case .cat:
-                imageView.image = UIImage(systemName: "cat")
-            case .verified:
-                imageView.image = UIImage(systemName: "checkmark.circle.fill")
-            case .locked:
-                imageView.image = UIImage(systemName: "lock.fill")
-            case .bot:
-                imageView.image = UIImage(systemName: "cpu")
-            default:
-                continue
-            }
-            
-            imageView.frame = CGRect(x: 0, y: 0, width: 16, height: 16)
-            markStackView.addArrangedSubview(imageView)
-        }
-        
-        // 更新更多按钮状态
-        moreButton.isHidden = userInfo.isMe
-        
-        // 开始流式布局，从关注/粉丝数下方开始
-        var currentY = followsCountLabel.frame.maxY + 10
-        
-        // 设置描述文本
-        if let description = userInfo.profile.description_?.markdown, !description.isEmpty {
-            let descriptionView = UIHostingController(
-                rootView: Markdown(description)
-                    .markdownInlineImageProvider(.emoji)
-            )
-            addSubview(descriptionView.view)
-            descriptionView.view.frame = CGRect(x: 16, y: currentY, width: frame.width - 32, height: 0)
-            descriptionView.view.sizeToFit()
-            currentY = descriptionView.view.frame.maxY + 16
-        }
-        
-        // 设置用户位置和URL
-        if let bottomContent = userInfo.profile.bottomContent {
-            switch onEnum(of: bottomContent) {
-            case .fields(let data):
-                let fieldsView = UserInfoFieldsView(fields: data.fields)
-                let hostingController = UIHostingController(rootView: fieldsView)
-                hostingController.view.frame = CGRect(x: 16, y: currentY, width: frame.width - 32, height: 0)
-                addSubview(hostingController.view)
-                hostingController.view.sizeToFit()
-                currentY = hostingController.view.frame.maxY + 16
-                
-            case .iconify(let data):
-                let stackView = UIStackView()
-                stackView.axis = .horizontal
-                stackView.spacing = 8
-                stackView.alignment = .center
-                
-                if let locationValue = data.items[.location] {
-                    let locationView = UIHostingController(
-                        rootView: Label(
-                            title: {
-                                Markdown(locationValue.markdown)
-                                    .font(.caption2)
-                                    .markdownInlineImageProvider(.emoji)
-                            },
-                            icon: {
-                                Image(uiImage: Asset.Image.Attributes.location.image
-                                    .withRenderingMode(.alwaysTemplate))
-                                    .imageScale(.small)
-                            }
-                        )
-                        .labelStyle(CompactLabelStyle())
-                        .padding(.horizontal, 8)
-                        .padding(.vertical, 4)
-                        .background(Color(.systemGray6))
-                        .cornerRadius(6)
-                    )
-                    locationView.view.sizeToFit()
-                    stackView.addArrangedSubview(locationView.view)
-                }
-                
-                if let urlValue = data.items[.url] {
-                    let urlView = UIHostingController(
-                        rootView: Label(
-                            title: {
-                                Markdown(urlValue.markdown)
-                                    .font(.caption2)
-                                    .markdownInlineImageProvider(.emoji)
-                            },
-                            icon: {
-                                Image(uiImage: Asset.Image.Attributes.globe.image
-                                    .withRenderingMode(.alwaysTemplate))
-                                    .imageScale(.small)
-                            }
-                        )
-                        .labelStyle(CompactLabelStyle())
-                        .padding(.horizontal, 8)
-                        .padding(.vertical, 4)
-                        .background(Color(.systemGray6))
-                        .cornerRadius(6)
-                    )
-                    urlView.view.sizeToFit()
-                    stackView.addArrangedSubview(urlView.view)
-                }
-                
-                if stackView.arrangedSubviews.count > 0 {
-                    addSubview(stackView)
-                    stackView.frame = CGRect(x: 16, y: currentY, width: frame.width - 32, height: 30)
-                    currentY = stackView.frame.maxY + 16
-                }
-            }
-        }
-        
-        // 更新视图总高度
-        frame.size.height = currentY
-    }
-    
-    private func setupDynamicBannerBackground(avatarUrl: String?) {
-        guard let avatarUrl = avatarUrl, let url = URL(string: avatarUrl) else { return }
-        
-        bannerImageView.kf.setImage(
-            with: url,
-            options: [
-                .transition(.fade(0.25)),
-                .processor(DownsamplingImageProcessor(size: CGSize(width: UIScreen.main.bounds.width * 2, height: 300))),
-                .scaleFactor(UIScreen.main.scale),
-                .cacheOriginalImage
-            ]
-        ) { [weak self] _ in
-            self?.blurEffectView.alpha = 0.7 // 增加模糊效果
-        }
-    }
-    
-    private func updateFollowButton(with userInfo: ProfileUserInfo) {
-        // 根据用户关系更新关注按钮状态
-        if userInfo.isMe {
-            followButton.isHidden = true
-        } else {
-            followButton.isHidden = false
-            if let relation = userInfo.relation {
-                let title = if relation.blocking {
-                    NSLocalizedString("profile_header_button_blockedblocked", comment: "")
-                } else if relation.following {
-                    NSLocalizedString("profile_header_button_following", comment: "")
-                } else if relation.hasPendingFollowRequestFromYou {
-                    NSLocalizedString("profile_header_button_requested", comment: "")
-                } else {
-                    NSLocalizedString("profile_header_button_follow", comment: "")
-                }
-                followButton.setTitle(title, for: .normal)
-            }
-        }
-    }
-    
-    @objc private func avatarTapped() {
-        onAvatarTap?()
-    }
-    
-    @objc private func bannerTapped() {
-        onBannerTap?()
-    }
-    
-    @objc private func followsCountTapped() {
-        onFollowsCountTap?()
-    }
-    
-    @objc private func fansCountTapped() {
-        onFansCountTap?()
-    }
-    
-    @objc private func moreMenuTapped() {
-        onMoreMenuTap?()
-    }
-}
